x86/perf/intel: Support adaptive PEBSv4
Adaptive PEBS is a new way to report PEBS sampling information. Instead of a
fixed size record for all PEBS events it allows to configure the PEBS record
to only include the information needed. Events can then opt in to use
such an extended record, or stay with a basic record which only contains
the IP.

The major new feature is support for including LBRs.
This allows (much faster) multi record "free running" PEBS, while still
supporting callstacks through callstack LBR.  So essentially a lot
of profiling can now be done without frequent interrupts, dropping
the overhead significantly.

The main requirement still is to use a period, and not use frequency mode,
because frequency mode requires reevaluating the frequency on each overflow.

Including floating point state (XMM) is also supported, which
allows efficient profiling of FP function arguments.

Adapt the drain function to handle variable length records.

Use a new callback to parse the new record format, and also
handle the STATUS field now being at a different offset.

Add code to set up the configuration register. Since there
is only a single register, all events either get the full
super set of all events, or only the basic record. We keep
track of this with separate bitmasks.

Currently 0day reports a ~5% regression on a workload sampled
with PEBS on a pre v4 CPU with this commit. So far not tracked down.

Signed-off-by: Andi Kleen <ak@linux.intel.com>