x86/perf/intel: Support PEBS on fixed counters
Icelake supports PEBS on fixed counters.

We need to change the order of PEBS and fixed counter enabling
to make sure PEBS is enabled for the fixed counters.

Also we need to handle adaptive PEBS correctly in this case,
by setting the right bits in the PEBS configuration register
on Icelake.

In addition the debug_store data structure has to be enlarged
to handle the additional reset fields for fixed counters.

Signed-off-by: Andi Kleen <ak@linux.intel.com>