FIXME: drm/i915/guc: disable guc lrc memcpy on engine reset
The golden context should be used only in case of reset handled
internally by GuC without kmd request or intervention (e.g. watchdog
reset). However, due to a bug in GuC firmware, the golden context is
being used to overwrite the guilty context in kmd-initiated engine
reset, which leads to issues. Since we don't currently use any feature
that requires GuC to handle resets internally, we can workaround the
bug by temporarily setting the golden context size to zero to make GuC
engine reset work while we wait for the fix to land in GuC firmware.

Note that this issue affects all firmwares gen11+, so we might need to
send a version of this workaround together with the ICL GuC enablement
if the firmware fix doesn't land in time. More fixes may also be
required if/when we start using features that require GuC to handle
internal resets.

v2: Trivial rebase

HSDES: 1406596847
Cc: Oscar Mateo <oscar.mateo@intel.com>
Cc: Michel Thierry <michel.thierry@intel.com>
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Acked-by: Oscar Mateo <oscar.mateo@intel.com>
Acked-by: Michel Thierry <michel.thierry@intel.com>
Signed-off-by: Oscar Mateo <oscar.mateo@intel.com>