x86/perf/intel: Add Icelake support
Add core Icelake core PMU perf code, including constraint tables and
the main enable code.

Icelake expanded the generic counters to always 8
even with HT on, but a range of events cannot be scheduled on the extra
4 counters. Add new constraint ranges to describe this to the scheduler.
The number of constraints that need to be checked is larger now than
with earlier CPUs.  At some point we may need a new data structure
to look them up more efficiently than with linear search.
So far it still seems to be acceptable however.

Icelake added a new fixed counter SLOTS. Don't warn for it. Full
support for it is added later in the patch series.

The cache events table is identical to Skylake

XXX double check with final events

Signed-off-by: Andi Kleen <ak@linux.intel.com>