INTERNAL [IOTG] drm: Don't assume that the primary plane always exists
In some virtualization use-cases, the operating system and hence
the graphics drivers may have to deal with a crtc(s) that does not
have a primary plane associated with it. For such cases, the
legacy non-atomic APIs should check for the existence of the primary
plane and return -EINVAL if it's not present.

v2(ssingh) : Ported this patch to 4.19 kernel.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Signed-off-by: Vivek Kasireddy <vivek.kasireddy@intel.com>
Change-Id: Ie16c4bdd19a3df25ca5c4241ded12b87ca5a106a