drm/i915/icl-11.5: Get the correct TBTPLL id
The enum definition for TBTPLL in 11.5 changes
to 3 which was earlier 2 on icl.
This has to be done to adhere to the current design
where pllids should match the shared pll array index.
Current design doesn't support addition of an new pll id in
the middle of the enum.

v2: Fix the return type of get_tbtpll_id function. (Madhav)
v3: Rebase (Lucas)

Reviewed-by: Madhav Chauhan <madhav.chauhan@intel.com>
Signed-off-by: Vandita Kulkarni <vandita.kulkarni@intel.com>
Acked-by: Lucas De Marchi <lucas.demarchi@intel.com>