ASoC: SOF: hw-apl: fix double free_irq
The error handling logic is incorrect when the DSP firmware is not
authenticated (e.g. if the CSME uses a different key)

In apl_probe, all the resources are supposed to be freed on error,
therefore on error cases it is needed to free the allocated irqs.

For all other routines called by the core, on error
snd_soc_sof_remove() will be called by the core. It is therefore not
needed to free the irq in apl_run_firmware

[    7.220237] Trying to free already-free IRQ 25
[    7.220245] ------------[ cut here ]------------
[    7.220253] WARNING: CPU: 0 PID: 3 at
/data/pbossart/ktest/upboard/kernel/irq/manage.c:1520 __free_irq+0x26f/0x2b0
[snip]
[    7.220383] Call Trace:
[    7.220390]  free_irq+0x38/0x70
[    7.220397]  apl_remove+0xc8/0x1c0 [snd_sof_intel_hda_common]
[    7.220403]  sof_probe+0x283/0x370 [snd_sof]

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>