FIXME: drm/i915/tgl: Gen12 csb support
The CSB format has been reworked for Gen12. Some of the events don't
have their own bit anymore and need to be inferred from other values.
To keep the logic simple, we convert the new gen12 format to the
gen11 format and re-use the legacy csb handling logic. This result
in a sliglthly slower handling for gen12 but allows us to get running
with a relatively small change and avoids code duplication. More
complex handling can be considered as a follow up.

FIXME: enable and test preempt-to-idle support after gen11 code is
merged.

v2: remove check on new_queue when setting ACTIVE_TO_IDLE, bug if
legacy_status is zero.

Cc: Michel Thierry <michel.thierry@intel.com>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Cc: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Acked-by: Michel Thierry <michel.thierry@intel.com>
Acked-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>