drm/i915/sim: Add the simulator placehoder and parameter.
Usually the simulator detection can be done according to HAS PCH
but if we are in a platform without PCH or if we force a different
PCI ID than simulator is expecting that auto detection fails.

So, let's introduce the placehoder but also include a way to force
is_simulator during boot.

v2 (James): Pay attention to is_simulator even when we can't already
detect we're in simulation

v3 (Lucas): Fix param checking and remove force to false since we never
really handle it. Values -1/1 were kept instead 0/1 to avoid breaking
existing users.

v4 (Daniele): add an IS_PRESILICON() macro and discourage direct
checking of the flag in dev_priv to prepare for upcoming replacement
with sim vs emu detection

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: James Ausmus <james.ausmus@intel.com>
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com> (v3)
Acked-by: James Ausmus <james.ausmus@intel.com>
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>