drm/i915/sim: silence an expected atomic error message
The vblank counters are completely crazy on simulation: they increment
by 2 every time you read them. I have confirmation that this part of
the simulator is not supposed to work as the real hardware, so silence
the error messages in order to allow us to better see the other, maybe
expected, error messages.

v2: Extend to TGL.
v3: Silence the vblank evade debug message too, it's also meaningless
in simulation. Leaving it as 11-12-only to allow other folks to
re-verify it later. We're expecting this to be fixed at some point
in the next 20 years.

v4 (Daniele): s/i915->is_simulator/IS_PRESILICON(i915)

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
Signed-off-by: Michel Thierry <michel.thierry@intel.com>
Cc: Lucas De Marchi <lucas.demarchi@intel.com>
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>