drm/i915/sim: Fulsim pci interface to mark pages as model owned
Pages that might be accessed concurrently by the simulated GT and the
CPU must be marked to be “model owned" (it means that access to this
page triggers a callback on HAS side and thanks to it we can inform
Fulsim that something was changed and that it may need to be processed).

v2:
  - Improve names (Michal)
  - BUG on invalid use (Michal)
  - SIM_HAS_ prefix to avoid confusion with HAS_ macros (Michal)
  - Remvoe newline (Michel)

v3: Rebased

v4: Use new mmio interface. This is a requirement for SRIOV because the
    VF can't communicate with HAS via PCI. The new interface is available
    starting from simics 7776 packages 4.8.pre348.1343 and 5.0.pre10.201

v5: s/i915->is_simulator/IS_PRESILICON(i915)

Signed-off-by: Michel Thierry <michel.thierry@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Michal Wajdeczko <michal.wajdeczko@intel.com>
Signed-off-by: Oscar Mateo <oscar.mateo@intel.com>
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>