drm/i915/sim: make sure we process all the display interrupts
What I'm seeing is that if we don't fully clear the IIR for display
interrupts we just stop getting more interrupts. We don't fully clear
the IIR due to the fact that we only call gen8_de_irq_handler() once,
which only writes zero to IIR once.

Now here's the big question: am I seeing this because it's a
Simics-related problem, or am I seeing this just because the simulator
is slower and implements less interrupts, giving us less opportunity
to actually accidentally clear the stalled IIRs? I'm thinking it could
actually be the second option, and if that's really the case we should
consider discarding this patch and loop-clearing all the IIRs on all
platforms...

v2: Rebased

v3 (Daniele): s/i915->is_simulator/IS_PRESILICON(i915)

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>