drm/i915/gvt: set digital bit in the virtual EDID
Sometimes physical monitor will not report the digital bit in its EDID,
which will cause GVT-g guest fail to enumerate the virtual HDMI monitor.
To avoid such issue, we will set the digital bit in virtual EDID we
present to guest.

This patch fixed the display issue of GVT-g guest on the APL NUC.

V2: Fixed the calculation of the sum.

Signed-off-by: Min He <min.he@intel.com>
Acked-by: Zhao Yakui <yakui.zhao@intel.com>