drm/i915/sim: hack port connected detection
Commit 'drm/i915/icl: implement the legacy HPD {dis,}connect flow for
HDMI' causes all ICL simulator HDMI ports being detect as disconnected.

It is a simulator problem, as it does not set the ICP_SDE_ISR/SDE_ISR as
it happens in real hardware. This simulator problem was never found
before because all simulator ports are HDMI, so if it was able to read
the edid of the port it was considered connected, what would fail in DP.

If the 'IS_ICELAKE()' is removed from the 'if' I was able to
reproduce this issue in all these simulators: ICP-HP, TGL, KBL, CFL
and CNL.

So here always returning port as connected for simulator, if
simulator is not emulating that port it it will fail when reading
EDID in HDMI and DP (if some day simulator started to emulate DP).

Cc: Paulo Zanoni <paulo.r.zanoni@intel.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>