NOT_UPSTREAM: drm/i915/tgl: HACK: disable advanced pre-fetch
Advanced pre-fetch restricts usage of self-modifying code, which we
use for the relocation emission. The correct approach would be to
disable the pre-fetcher before the relocation batch and re-enable it
afterwards, but it is slightly unclear how this is to be done: bspec
says that the pre-fetch must be disabled before the BBstart and
re-enabled as the first instruction in the batch, but one of the HW
engineers involved mentioned that the re-enabling will take effect
only on the next BBstart and recommended adding a dummy BBstart
to the next instruction.
While we wait for the HW team to confirm which is the best approach,
turn the feature off to avoid failures.

Bspec: 20947, 12609.
Cc: Kelvin Gardiner <kelvin.gardiner@intel.com>
Cc: Michel Thierry <michel.thierry@intel.com>
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Acked-by: Michel Thierry <michel.thierry@intel.com>