drm/i915/icl-11.5: allow display to have 4th pipe
We were previously assuming that number of transcoders is num_pipes + 1
for platforms with DDI. This is no longer true on ICL 11.5 since it
doesn't have a dedicated transcoder for eDP. Not taking the DSI
transcoders into account, num_pipes == num_transcoders.

Also add a warning to catch the cases in which this may change in future
instead of accessing garbage from the stack.

Cc: Mika Kahola <mika.kahola@intel.com>
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>