drm/i915/icl-11.5: Add hpd interrupt handling
ICP+ has 3 combo phys and 6 TC ports that are not entirely used on ICL,
where only 2 combos and 4 TC are used. Definitions for these registers
were kept with the ICL_ prefix in order to match the spec.

IRQ handling shared with ICP would add several branches for the platform
so instead this replicates the code with the different port
configuration and register bits/masks.

FIXME: once the code matures and ICL is on upstream we can revisit if
it's worth keeping it separate.

v2: change the approach to not share the array and paths with ICL in
order to a) avoid regression there; b) avoid rebase hell and c) make it
actually work as it should - this replaces the previous patch
"drm/i915/icl-11.5: add hpd pins for port C"

v3: rebase

Cc: Paulo Zanoni <paulo.r.zanoni@intel.com>
Cc: Mahesh Kumar <mahesh1.kumar@intel.com>
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Jose Souza <jose.souza@intel.com>