drm/i915/sim: Add force_pch param
Currently, every known pre-HW source of the PCH PCI ID is Intel Top
Secrest/Red Cover, and can thus not be allowed in the drm-intel-internal
repository. Add a module param to force the PCH type so we can avoid
plugging the PCH PCI IDs in, but still be able to support simulation
configurations that include the PCH model.

v2: Rebase (Lucas)

Suggested-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: James Ausmus <james.ausmus@intel.com>
Acked-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com> (v2)
Acked-by: James Ausmus <james.ausmus@intel.com>