LSM: Repair suppliers of inode_setxattr hooks
Both SELinux and Smack have errors in the way they
handle the inode_setxattr hooks. The infrastructure handles
cap_inode_setxattr, so that has to come out of the SELinux
and Smack hooks. They need to report -ENOSYS in the case
where the attribute in question isn't one they recognize.

Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>