net: stmmac: introduce IEEE 802.1Qbu implementation
The following functionalities are added:-

PEC is meant to specify which TxQ is preemptable queue. By default,
TxQ0 must NOT be express queue.

AFSZ is meant to change the non-final fragment size of preempted
frame. A smaller value here means express frame will be transmitted by
express MAC as soon as preempted MAC (pMAC) has collected the configured
fragment size, default to 64-byte fragment.

HADV is meant to specify the time (in nano-seconds) for pMAC to
switch from release to hold state in advance.

RADV is meant to specify the time (in nano-seconds) for pMAC to
switch from hold to release state in advance.

pMAC STS is to indicate whether preemptable MAC is in either Hold
or Release state.

The newly added IEEE 802.1Qbu functionalities are also hooked into
the driver main logics:
 a) stmmac_set_features() & stmmac_dvr_probe(): to independently
    enable/disable FPE functionality
 b) stmmac_hw_setup: to include fprq (frame preemption residue queue)
    setting.

Signed-off-by: Ong Boon Leong <boon.leong.ong@intel.com>