Kernel/drm/i915: Remove the redundant cache change for shared_page
The gvt-g will try to map one shared_page for PVMMIO usage so that the
guest system can pass the mmio access through PVMMIO page. Currently
it will firstly try to map the page and then change the cache type.
In fact as the shared_page is accessed through MMIO, it is meaningless
to change the cache type.
So it will remove the redundant cache change for the shared_page.

Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>
Signed-off-by: Li, Fei1 <fei1.li@intel.com>
Reviewed-by: Eddie Dong <eddie.dong@intel.com>
Reviewed-by: Jiang Fei <fei.jiang@intel.com>