vhm: adapt to the new state transition of VHM requests
This is the counter-part of VHM request state transition update.

Instead of using two members (namely ''valid'' and ''processed''), the new state
transition uses a single member (i.e. ''processed) following the transition
pattern below.

       FREE -> PENDING -> PROCESSING -> COMPLETE -> FREE -> ...

The ownership of a VHM request shifts according to its state. Namely:

1) When a request is in COMPLETE or FREE state, the request is owned by the
   hypervisor. SOS (VHM or DM) shall not read or write the internals of the
   request except the state.

2) When a request is in PENDING or PROCESSING state, the request is owned by
   SOS. The hypervisor shall not read or write the request other than the state.

Atomic operations should be used to access the state. For this the ''processed''
member in vhm_request is typed 'atomic_t' which is essentially the same 32-bit
signed integer. This allows us to use atomic operations on the state directly,
instead of casting it everywhere.

It is also worth noting that atomic_set() in Linux does not imply a memory
fence; it only includes a barrier to prevent compiler reordering. Thus smp_mb()
are added before changing the state to COMPLETE so that the hypervisor always
has the right results after it sees the state change.

v1 -> v2:

    * Keep the original flow (i.e. update state after a client is found) instead
      of atomic_cmpxchg, given that acrn_ioreq_distribute_request() is called
      in tasklet and only on cpu 0.
    * Add documentation on struct vhm_request, following the kernel-doc style.

Signed-off-by: Junjie Mao <junjie.mao@intel.com>
Reviewed-by: Zhao Yakui <yakui.zhao@intel.com>