drm/i915/gvt: Check the state of PVMMIO gtt table to avoid incorrect calling for hypervisor
Now hypervisor adds strict checks when the sos tries to add/delete the mapping
between gfn and mfn. In such case it should check the state of PVMMIO gtt
table to avoid the incorrect hyper call. If the mapping between gfn and mfn
is not setup for PVMMIO GTT, it should not remove the mapping. Otherwise
the hypervisor will report that the mapping doesn't exist.

Tracked-On: https://github.com/projectacrn/acrn-hypervisor/issues/1158
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>
Reviewed-by: He, Min <min.he@intel.com>