drm/i915/gvt: Fixed an issue in PS MMIO handlers
When GVT-g guest tries to disable a scaler, it will write 0 to
PS_CTRL register. However, in such case, GVT-g PS MMIO handler will
not allow this value to be written to the HW, which will cause scaler
cannot be disabled.

This patch fixed the issue mentioned above.

V2: Uses PS_SCALER_EN to check if scaler is enabled.

Tracked-on: https://github.com/projectacrn/acrn-hypervisor/issues/1188
Signed-off-by: Min He <min.he@intel.com>
Signed-off-by: Fei Jiang <fei.jiang@intel.com>
Reviewed-by: Zhao Yakui <yakui.zhao@intel.com>