Kernel/VHM: Use the traped pci_access addr based on VM instead of static global variable
The emulation of PCI access is divided into two steps:
1. write the access addr into 0xCF8((bus << 16) | (dev << 11) | (func << 8) | (reg))
2. read/write the 0xCFC port

Now the static variable is defined to obtain the cached_bus/dev/func/reg.

But if the pci emulated access is triggered on two VM, it will cause that
the incorrect bus/dev/func/reg is cached.
So it is emulated based on per-vm instead of static global variable.

At the same time the pci_ext space is also supported. It is defined in
((reg & 0xF00) << 16).

Tracked-On: https://github.com/projectacrn/acrn-hypervisor/issues/1957
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>
Reviewed-by: Eddie Dong <eddie.dong@intel.com>
Tested-by: Yin, FengWei <fengwei.yin@intel.com>