drivers/hyper_dmabuf: Reuse the exported dma_buf to avoid creating multiple dma_buf for one importer
After the imported hyper_dmabuf is added to the BE, the
IOCTL_HYPER_DMABUF_EXPORT_REMOTE iotcl will export one dma_buf based on
the imported ID info. When the ioctl is called twice for the same ID,
it will try to create two dma_buf structure and then sync with the FE
driver. In fact one dma_buf is enough for the same imported ID info. And
the different file_fd is returned for the same dma_buf when the ioctl
is called twice.

Tracked-On: projectacrn/acrn-hypervisor#2237
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>
Reviewed-by: Liu, Shuo <shuo.a.liu@intel.com>
Acked-by: Dongwon Kim <dongwon.kim@intel.com>