Kernel/VHM: Pass ioreq_client as ioreq_client_thread argument to avoid dead loop
Sometimes the ioreq_client_thread has no opportunity to be executed after
it is successfully created. Then when the ioreq_client is destroyed, the
thread is scheduled and internal state of ioreq_client can't be updated
correctly as acrn_ioreq_get_client returns NULL for destroyed client_id.
In such case it needs to wait for the state to release the resource and
dead loop is detected.
So the ioreq_client is used as the parameter of ioreq_client_thread and
assure that the internal state can be updated.

V2-V3: Consider the length limitation of kthread name and refine
the passed argument of kthread_run when creating the kthread.

Tracked-On: projectacrn/acrn-hypervisor#2383
Reviewed-by: He Min <min.he@intel.com>
Reviewed-by: Shuo A Liu <shuo.a.liu@intel.com>
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>