kernel/vhm: Use kthread API to simplify the lifecycle of ioreq_client_thread
Now the complex state is updated in ioreq_client for ioreq_client_thread
(For example: IOREQ_CLIENT_EXIT)
In fact Linux provides the kthread API, which can be used to manage the
lifecycle of ioeq_client_thread.
When it needs to be destroyed, the kthread_stop is used to wake up and
terminate the ioreq_client_thread. And the kthread_should_stop is used
in the ioreq_client_thread.

V2-V3: Remove the acrn_ioreq_put_client in ioreq_client_thread.
Instead it is always acquired when ioreq_client_thread is alive.
Otherwise the refcnt of ioreq_client is incorrect if the thread
is created while it is not executed.

Tracked-On: projectacrn/acrn-hypervisor#2383
Reviewed-by: He Min <min.he@intel.com>
Signed-off-by: Shuo Liu <shuo.a.liu@intel.com>
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>