drm/i915/gvt: GVT uses kernel thread mechanism of ioreq_client instead of specific handle
The GVT will create the ioreq_client and then create the specific thread
to handle the emulated io_req. As the ioreq_client provides one thread
model to handle the emulated io_req, it is unnecessary for acrn_gvt to
create the specific kthread to handle the emulated io_req.
In such case all the kernel thread related with ioreq_client can be
handled by using the same style.

V2->V3: Change the acrngt_handle_kick to acrngt_emulate_ioreq
as it doesn't use the VBS-K kick function

Tracked-On: projectacrn/acrn-hypervisor#2383
Reviewed-by: Shuo A Liu <shuo.a.liu@intel.com>
Reviewed-by: He Min <min.he@intel.com>
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>