drm/i915/gvt: limit the active perf on BXT platform
In commit "drm/i915/gvt: force to active the high-performance mode
during vGPU busy", it set the max GPU freq when there's GVT workload,
but on BXT platform, the max GPU freq will impact the CPU performance,
To make a balance between GPU and CPU, we hardcode it to 600Mhz on BXT
platform.
Also, this patch will not disable rps interrupt anymore, so that if
workload is heavier, GPU freq can be adjusted to higher.

Tracked-On: projectacrn/acrn-hypervisor#2537
Signed-off-by: Min He <min.he@intel.com>
Reviewed-by: Zhao Yakui <yakui.zhao@intel.com>