Kernel/VHM: Destroy ioreq_client in ioeventfd_deinit instead of put_ioeventfd_info
Currently the ioreq_client of ioeventfd_info list is destroyed only when
the ioevent_info needs to be released. It is released too late.
So when the acrn_ioeventfd_deinit is called, it will firstly try to
release the ioreq_client and then release the other resources related
with ioevent_info list.

Tracked-On: projectacrn/acrn-hypervisor#2383
Reviewed-by: Shuo A Liu <shuo.a.liu@intel.com>
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>