drm/i915/gvt: defer the release of gvt iorq_client to workaround the issue of adding/removing iorange on invalid ioreq_client
Currently acrngt_instance_destroy will firstly try to destroy the gvt
ioreq_client and then continue the sequence of releasing vgpu resource.

But as the vgpu_release calls acrn_ioreq_del_iorange to remove the trapped
emulated ioreq range, the error is returned by acrn_ioreq_del_iorange and
then the vgpu_release doesn't execute the cleanup correctly.
So the gvt ioreq_client will be deferred to be released so that it can
workaround the issue of adding/removing iorange on invalid ioreq_client.

Tracked-On: projectacrn/acrn-hypervisor#2440
Reviewed-by: Shuo A Liu <shuo.a.liu@intel.com>
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>