kernel/vhm: acrn_ioreq_del_range doesn't return error for the invalid ioreq_client
Currently the acrn_ioreq_destroy_client will release the added trapped
iorange so that the corresponding resource still can be released even
when the ioreq_client user is terminated with some exception.

Under multi-thread scenario maybe the ioreq_client is already released
while acrn_ioreq_del_iorange is still called to detach the emulated
ioreq range. In such case it will fail on checking the returned value.
So it won't return error any more when the ioreq_client is invalid.

Tracked-On: projectacrn/acrn-hypervisor#2440
Reviewed-by: Shuo A Liu <shuo.a.liu@intel.com>
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>