drm/i915/gvt: Remove the redundant range op on gvt ioreq_client for PVMMIO
The emulated io range is added for ioreq_client so that the VHM can
distribute the trapped ioreq to the corresponding ioreq_client.
Now the gvt ioreq_cleint already adds the PCI bar 0 to the trapped range.
Later it will add PVMMIO range  to the trapped range for gvt ioreq_client
when PVMMIO is enabled. While the PVMMIO is disabled, it will add back
the trapped range. It is redundant.

Tracked-On: projectacrn/acrn-hypervisor#2440
Reviewed-by: Shuo A Liu <shuo.a.liu@intel.com>
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>