acrn/vhm: Use the mempool to allocate the buffer for hypercall parameter
Currently the buf for hypercall parameter resides on the kernel statck
and then the virt_to_phys is used to get the guest physical addr.
But it can't work when the VMAP_STACK is enabled as the virt_to_phys can't
return the correct physical address if the stack is allocated by using
vmalloc.
Use the kmalloc to allocate the required buffer for hypercall and then use
virt_to_phys to get the physical address. The physical address is used as
the parameter of copy_from/to_gpa in hypervisor.
To avoid the failure of kmalloc, the mempool is used.

Tracked-On: projectacrn/acrn-hypervisor#1318
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>
Reviewed-by: Yin Fengwei <fengwei.yin@intel.com>