acrn/vhm: Change some stack variable of hypercall parameter as static
When loading acrn trace/hv_log module, it needs to call the hcall_get_hw_info
to query the cpu info in hypervisor. And the physical address of stack variable
is passed as the parameter of hcall_get_hw_info. This can't work with VMAP_STACK
enabled.
At the same time as it is used only once in the initialization stage, it is
changed as the static. In such case the virt_to_phys still can work and
it can avoid the allocation failure.
Similiarly the hc_api_version is also changed as static efore initializing
acrn memory-pool.

v1->v2: Remove the unnecessary memset for static definition
v2->v3: Use the slow_virt_to_phys to get the GPA of static variable in
trace/hvlog module

Tracked-On: projectacrn/acrn-hypervisor#1318
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>
Reviewed-by: Yin Fengwei <fengwei.yin@intel.com>