acrn/vhm: HC_ASSIGN_PTDEV passes the bdf directly instead of guest physical address
Currently the HC_ASSIGN/DEASSIGN_PTDEV is called to add/remove the
pass-through device and the bdf is passed by using guest_physical_addr
mode. This requires that the ACRN hypervisor looks up the GPA mapping
to get the bdf. In fact it looks much simpler that it is passed directly.

Tracked-On: projectacrn/acrn-hypervisor#1318
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>
Reviewed-by: Yin Fengwei <fengwei.yin@intel.com>
Tracked-On: PKT-2115