drm/i915/gvt: enable local direct display for WaaG
align the display type with the one in the virtual vbt and add missing
handler for sprite plane.

Guest OS (Windows 10, eg) which has no idea of the plane restriction can
output to a local monitor with a specific service OS now. The SOS should
enable the plane restriction feature and leaves the first pipes for
the guest OS.

v1: separated from the previous patch enabling opregion

Tracked-On: projectacrn/acrn-hypervisor#3106
Signed-off-by: Liu Xinyun <xinyun.liu@intel.com>
Reviewed-by: Zhao Yakui <yakui.zhao@intel.com>