EDAC, igen6: Add registration APIs for In-Band ECC error notification
There are some external modules which want to be notified about the
In-Band ECC error for specific error handling. Because the In-Band ECC
error event is captured by the igen6_edac driver via NMI, so add the
registration APIs for the In-Band ECC error event that those modules
can be notified for further error handling.

Signed-off-by: Qiuxu Zhuo <qiuxu.zhuo@intel.com>