sep: Change some stack variable of hypercall parameter as static
The UTILITY_Read_Cpuid needs to call the acrn_hypercall2 directly
to query the cpuid_info in hypervisor. And the physical address of stack variable
is passed as the parameter . This can't work with VMAP_STACK enabled.
At the same time as it is used only once in the initialization stage, it is
changed as the static. In such case the virt_to_phys still can work and
it can avoid the allocation failure.

Tracked-On: projectacrn/acrn-hypervisor#1318
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>
Reviewed-by: Yin Fengwei <fengwei.yin@intel.com>