Change the priority of acrngt_emulation_thread on demand when injecting VBLANK interrupt
When pipe vblank interrupt is injected, the priority of
acrngt_emulation_thread will be increased so that it can handle the
emulated i915 trap in time. It will be restored after 750us, which is
handled in one higher priority RT thread.(It needs to wait for the timer
and then change the priority of acrngt_emulation_thread to normal).
The RT is used when injecting the Vblank interrupt.

V1-V2: Preempt is disabled when changing the priority and starting timer
in function of injecting vblank interrupt for UOS

Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>
Signed-off-by: Liu Shuo <shuo.liu@intel.com>