x86/resctrl: Enable L3 CAT when platform 0x8C supports it
Some 0x8C platforms do support L3 CAT but when this is the case
there is no CPUID enumeration to indicate this to the OS.

To support L3 CAT on the 0x8C platforms that support this feature
it is necessary to manually probe for this feature and hardcode the
properties that would have been provided if CPUID enumeration was
supported. When doing so take care to ensure that this is not used
instead of CPUID enumeration, which is still used for the L2 CAT feature
supported by these platforms.

Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>