i915/gvt: obtain power domain to take the snapshot of gvt registers correctly
When the VGPU workload is submitted, it needs to check whether the some
vgpu registers are consistent with that in gvt registers. The gvt register
is captured without holding the corresponding power domain. In such case
the snapshot of some registers are incorrect and it complains the
unconformance mmio access on VGPU.

Tracked-On: projectacrn/acrn-hypervisor#3520
Reviewed-by: He Min <min.he@intel.com>
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>